FASTAPI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.206
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
python-multipart==0.0.6
pydantic==2.4.2
